fix: showing errors added
