+++ src/components/Login/RegisterForm.js
@@ -8,8 +8,11 @@ import { useDispatch, useSelector } from "react-redux";
 import { authActions } from "../../store/auth";
 import MainButton from "../Reusable/MainButton";
 import PasswordInput from "../Reusable/PasswordInput";
+import { signupFunction } from "../../store/auth-actions";
+import SmallSpinner from "../Reusable/SmallSpinner";
 
 const LoginForm = () => {
+  const showSpinner = useSelector((state) => state.spinner.show);
   const history = useHistory();
   const dispatch = useDispatch();
   const isError = useSelector((state) => state.auth.error);
@@ -50,35 +53,7 @@ const LoginForm = () => {
   const submitHandler = (event) => {
     event.preventDefault();
 
-    const sendRequest = async () => {
-      const response = await fetch("/api/v1/register", {
-        method: "POST",
-        body: JSON.stringify({
-          email: enteredEmail,
-          password: enteredPassword,
-        }),
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-
-      if (!response.ok) {
-        throw new Error("This email already exists");
-      }
-
-      const data = await response.json();
-      setRegComplete(true);
-      setTimeout(() => {
-        setRegComplete(false);
-        dispatch(authActions.logIn(data.access_token));
-      }, 2000);
-    };
-
-    try {
-      sendRequest();
-    } catch (err) {
-      dispatch(authActions.showError(err.message));
-    }
+    dispatch(signupFunction(enteredEmail, enteredPassword, setRegComplete));
 
     document.activeElement.blur();
     resetEmailInput();
@@ -94,57 +69,64 @@ const LoginForm = () => {
     emailInputHasError || isError ? `formInput wrongCredentials` : `formInput`;
 
   return (
-    <div className={classes.container}>
-      <form onSubmit={submitHandler} className={classes.loginForm}>
-        {!regComplete && (
-          <>
-            <h1>Sign Up</h1>
-            <div className={classes.newUser}>
-              <p>Have an account?</p>
-              <span onClick={signInHandler}>Sign in.</span>
-            </div>
-            <div className="inputsContainer">
-              <input
-                type="email"
-                placeholder="Email"
-                className={emailInputClasses}
-                value={enteredEmail}
-                onChange={emailChangedHandler}
-                onBlur={emailBlurHandler}
-                onFocus={emailFocusHandler}
-              ></input>
-              <PasswordInput
-                inputType="password"
-                enteredPassword={enteredPassword}
-                passwordInputHasError={passwordInputHasError}
-                enteredPasswordIsValid={enteredPasswordIsValid}
-                passwordChangedHandler={passwordChangedHandler}
-                passwordBlurHandler={passwordBlurHandler}
-                resetPasswordInput={resetPasswordInput}
-                placeholder="Password"
+    <>
+      {showSpinner && <SmallSpinner />}
+      <div className={classes.container}>
+        <form onSubmit={submitHandler} className={classes.loginForm}>
+          {!regComplete && (
+            <>
+              <h1>Sign Up</h1>
+              <div className={classes.newUser}>
+                <p>Have an account?</p>
+                <span onClick={signInHandler}>Sign in.</span>
+              </div>
+              <div className="inputsContainer">
+                <input
+                  type="email"
+                  placeholder="Email"
+                  className={emailInputClasses}
+                  value={enteredEmail}
+                  onChange={emailChangedHandler}
+                  onBlur={emailBlurHandler}
+                  onFocus={emailFocusHandler}
+                ></input>
+                <PasswordInput
+                  inputType="password"
+                  enteredPassword={enteredPassword}
+                  passwordInputHasError={passwordInputHasError}
+                  enteredPasswordIsValid={enteredPasswordIsValid}
+                  passwordChangedHandler={passwordChangedHandler}
+                  passwordBlurHandler={passwordBlurHandler}
+                  resetPasswordInput={resetPasswordInput}
+                  placeholder="Password"
+                />
+              </div>
+              {validationError && (
+                <p className={classes.errorText}>
+                  Please enter correct email & password.
+                </p>
+              )}
+              {isError && (
+                <p className={classes.errorText}>{isError.errorMessage}</p>
+              )}
+              <MainButton
+                title="Sign Up"
+                type="submit"
+                disabled={!formIsValid}
               />
-            </div>
-            {validationError && (
-              <p className={classes.errorText}>
-                Please enter correct email & password.
-              </p>
-            )}
-            {isError && (
-              <p className={classes.errorText}>{isError.errorMessage}</p>
-            )}
-            <MainButton title="Sign Up" type="submit" disabled={!formIsValid} />
-          </>
-        )}
-        {regComplete && (
-          <span className={classes.sucessReg}>
-            Congratulations!
-            <br />
-            You have successfully registered!
-          </span>
-        )}
-      </form>
-      <img src={logoMain} alt="keysee" className={classes.mainLogo} />
-    </div>
+            </>
+          )}
+          {regComplete && (
+            <span className={classes.sucessReg}>
+              Congratulations!
+              <br />
+              You have been successfully registered!
+            </span>
+          )}
+        </form>
+        <img src={logoMain} alt="keysee" className={classes.mainLogo} />
+      </div>
+    </>
   );
 };
